RENAMING hand_manipulation_suite to hands
