+++ robohive/tests/test_all.py
@@ -1,7 +1,7 @@
 import unittest
 from robohive.tests.test_arms import TestArms
 from robohive.tests.test_fm import TestFM
-from robohive.tests.test_hand_manipulation_suite import TestHMS
+from robohive.tests.test_hands import TestHMS
 from robohive.tests.test_multitask import TestKitchen
 from robohive.tests.test_myo import TestMyo
 from robohive.tests.test_claws import TestClaws
